refactor: drop shred `Query` namespace
