+++ extensions/tags/src/Api/Controller/ListTagsController.php
@@ -12,7 +12,7 @@ namespace Flarum\Tags\Api\Controller;
 use Flarum\Api\Controller\AbstractListController;
 use Flarum\Http\RequestUtil;
 use Flarum\Http\UrlGenerator;
-use Flarum\Query\QueryCriteria;
+use Flarum\Search\SearchCriteria;
 use Flarum\Tags\Api\Serializer\TagSerializer;
 use Flarum\Tags\Search\TagSearcher;
 use Flarum\Tags\TagRepository;
@@ -53,7 +53,7 @@ class ListTagsController extends AbstractListController
         }
 
         if (array_key_exists('q', $filters)) {
-            $results = $this->searcher->search(new QueryCriteria($actor, $filters), $limit, $offset);
+            $results = $this->searcher->search(new SearchCriteria($actor, $filters), $limit, $offset);
             $tags = $results->getResults();
 
             $document->addPaginationLinks(
