+++ framework/core/src/Api/Controller/ListAccessTokensController.php
@@ -13,7 +13,7 @@ use Flarum\Api\Serializer\AccessTokenSerializer;
 use Flarum\Http\Filter\AccessTokenSearcher;
 use Flarum\Http\RequestUtil;
 use Flarum\Http\UrlGenerator;
-use Flarum\Query\QueryCriteria;
+use Flarum\Search\SearchCriteria;
 use Psr\Http\Message\ServerRequestInterface;
 use Tobscure\JsonApi\Document;
 
@@ -37,7 +37,7 @@ class ListAccessTokensController extends AbstractListController
         $limit = $this->extractLimit($request);
         $filter = $this->extractFilter($request);
 
-        $tokens = $this->searcher->search(new QueryCriteria($actor, $filter), $limit, $offset);
+        $tokens = $this->searcher->search(new SearchCriteria($actor, $filter), $limit, $offset);
 
         $document->addPaginationLinks(
             $this->url->to('api')->route('access-tokens.index'),
