+++ framework/core/src/Api/Controller/ListGroupsController.php
@@ -13,7 +13,7 @@ use Flarum\Api\Serializer\GroupSerializer;
 use Flarum\Group\Filter\GroupSearcher;
 use Flarum\Http\RequestUtil;
 use Flarum\Http\UrlGenerator;
-use Flarum\Query\QueryCriteria;
+use Flarum\Search\SearchCriteria;
 use Psr\Http\Message\ServerRequestInterface;
 use Tobscure\JsonApi\Document;
 
@@ -42,7 +42,7 @@ class ListGroupsController extends AbstractListController
         $limit = $this->extractLimit($request);
         $offset = $this->extractOffset($request);
 
-        $criteria = new QueryCriteria($actor, $filters, $sort, $sortIsDefault);
+        $criteria = new SearchCriteria($actor, $filters, $sort, $sortIsDefault);
 
         $queryResults = $this->searcher->search($criteria, $limit, $offset);
