+++ framework/core/src/Search/SearchCriteria.php
@@ -7,7 +7,7 @@
  * LICENSE file that was distributed with this source code.
  */
 
-namespace Flarum\Query;
+namespace Flarum\Search;
 
 use Flarum\User\User;
 
@@ -16,7 +16,7 @@ use Flarum\User\User;
  * query. The limit and offset are not included because they only determine
  * which part of the entire result set will be returned.
  */
-class QueryCriteria
+class SearchCriteria
 {
     public function __construct(
         public User $actor,
